Split tests related to options parsing
